Add updates from Cryomundus
It seems some of the tweaks done on the original repo didn't contain those that I had when I grabbed the modified PK3 from the discord, so some merging has been done here.  Overall the gun feels pretty good and retains some of the "swap hands" and other mechanics, while also using the .50 AE to .50 AM swap that was done.