Draw Scope Logic Cleanup
Mostly to try and prevent a divide by 0 error when it sets deg to 1/degree, which somehow could resolve into a 0?  Either way the rest of the math here is a combination of the original values as well as some taken from the Brontornis Rifle instead of the Boss Rifle.